requirements.txt
streamlit==0.71.0
Keras
numpy
opencv_python==4.4.0.46
tensorflow-cpu
matplotlib